langgraph/requirements.txt
streamlit
langchain-groq
langchain
python-dotenv
langchain_community
langgraph
IPython
beautifulsoup4
tiktoken
sentence-transformers
chromadb
selenium
absl-py
astunparse
tf-keras
sentence-transformers